Rename the context_dir file to make it more consistent.
